Update subspacecomp.g
filter: after the space has been computed, loop through all g(e^n) cells and verify that they are all homotopy equivalent to an (n-1)-ball. this will be computationally time consuming so ensure that it's done only once for a given space and then store its compatibility with the algorithm as a boolean component object.

to do: include those cells in the boundary of the output target space to the source space. the desired output will be an inclusion from the boundary of the complement of a thickened subspace to that complement in its entirety